+++ quickcheck-plutus-instances/CHANGELOG.md
@@ -9,13 +9,17 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 ### Added
 
 * Instances of `Arbitrary`, `CoArbitrary` and `Function` for
-  `Plutus.V1.Value.Value`.
+  `Value`, `Address`, `Credential`, `TxOut`, `StakingCredential`. 
 * Instances of `Arbitrary`, `Arbitrary1`, `Arbitrary2`, `CoArbitrary` and
   `Function` for `PlutusTx.AssocMap.Map`.
+* `UniqueMap` wrapper around `AssocMap.Map` which ensures unique keys.
+* `UniqueList` wrapper around lists to ensure unique values.
+* `NonNegative` wrapper for generating non-negative (Plutus) numerical values.
 
 ### Changed
 
 * Fix instances for `TxOutRef` so that they respect invariants.
+* Fix instances for `Value` to use `UniqueMap` to respect invariants.
 
 ## 1.0 -- 2021-11-15